+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The following changes have been implemented but not released yet:
 
 ## [Unreleased]
 
+### New features
+
+- `getWebIdProfileAll` and `getWebIdProfileAll`: functions to discover one's personal
+  profile (which may be a Solid resource) if discoverable from their WebID profile
+  document (which may NOT be a Solid resource).
+
 - Allow 'createContainerAt()' to take an optional SolidDataset parameter to
   use as the body of the HTTP request to the server. This is really useful
   when we wish to include meta-data for a new container, things like a textual